+++ lib/src/nb_asset_tracking_flutter_method_channel.dart
@@ -231,4 +231,9 @@ class MethodChannelNbAssetTrackingFlutter
     return await methodChannel.invokeMethod("isTripInProgress");
   }
 
+  @override
+  Future<String> setupUserId({required String userId}) async {
+    return await methodChannel.invokeMethod("setupUserId", userId);
+  }
+
 }